+++ src/errors/map/InvalidChunkDataError.ts
@@ -1,5 +1,5 @@
 export default class InvalidChunkDataError extends Error {
     constructor() {
-        super("Invalid Chunk Data input.");
+        super("Invalid chunk data input.");
     }
 }