+++ src/errors/map/InvalidTileDataError.ts
@@ -1,0 +1,5 @@
+export default class InvalidTileDataError extends Error {
+    constructor() {
+        super("Invalid tile data input.");
+    }
+}